Fully replace market
